replace mal with jikan and update test database
